feat: slicer tool modes (pointer/knife), boundary drag/delete, progress bar; fix MSVC warnings
- Add Pointer tool (V): select/drag boundaries, right-click to delete
- Add Knife tool (C): left-click to add new slice lines
- Enable progress bar in AudioFileListPanel for labeling progress
- Add boundary selection highlight and drag preview in WaveformDisplay
- Right-click empty area plays the surrounding slice segment instantly
- Fix C4068: wrap GCC pragmas with #ifdef __GNUC__ in test files
- Fix C4244: add static_cast<int> for int64_t narrowing in Slicer.cpp

+++ src/infer/audio-util/src/Slicer.cpp
@@ -110,7 +110,7 @@ namespace AudioUtil
 
         s.m_sampleRate = sampleRate;
         s.m_threshold = std::pow(10, params.threshold / 20.0);
-        s.m_hopSize = divIntRound<int64_t>(static_cast<int64_t>(params.hopSize) * sampleRate, 1000LL);
+        s.m_hopSize = static_cast<int>(divIntRound<int64_t>(static_cast<int64_t>(params.hopSize) * sampleRate, 1000LL));
         s.m_winSize = static_cast<int>(std::min(
             divIntRound<int64_t>(static_cast<int64_t>(params.minInterval) * sampleRate, 1000LL),
             static_cast<int64_t>(4) * s.m_hopSize));
